set local env / initial secret bucket

+++ devops/environment/base/setup.sh
@@ -3,7 +3,10 @@
 project=${COMPOSE_PROJECT_NAME:?}
 app_dir=${APP_DIR:?}
 staging_env=${STAGING_ENV:?}
+json="${app_dir}/devops/bin/json.sh"
+openssl="${app_dir}/devops/bin/openssl.sh"
 
+# docker infrastructure
 docker build --force-rm \
     --build-arg "archive=${staging_env}-current" \
     --tag "${project}/archive" \
@@ -18,15 +21,20 @@ docker build --force-rm \
     "${app_dir}/devops/docker/php" \
 ;
 
+# file management
 cd files
 
-case "${staging_env}" in
-    dev)
-        if [ ! -f mhsendmail ]; then
-            curl -sS -o mhsendmail --fail -L https://github.com/mailhog/mhsendmail/releases/download/v0.2.0/mhsendmail_linux_amd64
-            [ $? -ne 0 ] && cd - && exit 1
-        fi
-        ;;
-esac
+if [ "${staging_env}" = 'dev' ] && [ ! -f mhsendmail ]; then
+    curl -sS -o mhsendmail --fail -L https://github.com/mailhog/mhsendmail/releases/download/v0.2.0/mhsendmail_linux_amd64
+    [ $? -ne 0 ] && cd - && exit 1
+fi
+
+cd - >/dev/null
+
+# secret management
+cd "../${staging_env}/secrets"
+
+${json} -f bucket.json '{"APP_SECRET": "boo"}'
+[ $? -ne 0 ] && cd - && exit 1
 
 cd - >/dev/null
